+++ test/util.js
@@ -81,3 +81,15 @@ exports.todo = function (test, fn) {
     // add more as needed
   });
 };
+
+exports.injectIFrameWithScript = function (document, scriptStr) {
+  scriptStr = scriptStr || "";
+  let iframe = document.createElement('iframe');
+  document.body.appendChild(iframe);
+
+  let scriptTag = iframe.contentWindow.document.createElement('script');
+  scriptTag.innerHTML = scriptStr;
+  iframe.contentWindow.document.body.appendChild(scriptTag);
+
+  return iframe;
+};
